Updated the model type to more accurately reflect a price.

+++ product/models.py
@@ -4,7 +4,7 @@ from django.db import models
 # Create your models here.
 class Product(models.Model):
     title = models.CharField(max_length=150)
-    price = models.IntegerField()
+    price = models.DecimalField(decimal_places=2, max_digits=5)
     description = models.TextField()
 
     def __str__(self):
